+++ packages/block-editor/src/components/colors-gradients/control.js
@@ -141,15 +141,15 @@ function ColorGradientControlInner( {
 								}
 							>
 								<Tabs.TabList>
-									<Tabs.Tab id={ TAB_IDS.color }>
+									<Tabs.Tab tabId={ TAB_IDS.color }>
 										{ __( 'Solid' ) }
 									</Tabs.Tab>
-									<Tabs.Tab id={ TAB_IDS.gradient }>
+									<Tabs.Tab tabId={ TAB_IDS.gradient }>
 										{ __( 'Gradient' ) }
 									</Tabs.Tab>
 								</Tabs.TabList>
 								<Tabs.TabPanel
-									id={ TAB_IDS.color }
+									tabId={ TAB_IDS.color }
 									className={
 										'block-editor-color-gradient-control__panel'
 									}
@@ -158,7 +158,7 @@ function ColorGradientControlInner( {
 									{ tabPanels.color }
 								</Tabs.TabPanel>
 								<Tabs.TabPanel
-									id={ TAB_IDS.gradient }
+									tabId={ TAB_IDS.gradient }
 									className={
 										'block-editor-color-gradient-control__panel'
 									}