+++ packages/components/CHANGELOG.md
@@ -22,6 +22,10 @@
 
 -   `DropdownMenuV2Ariakit`: prevent prefix collapsing if all radios or checkboxes are unselected ([#56720](https://github.com/WordPress/gutenberg/pull/56720)).
 
+### Experimental
+
+-   `Tabs`: implement new `tabId` prop ([#56883](https://github.com/WordPress/gutenberg/pull/56883)).
+
 ## 25.13.0 (2023-11-29)
 
 ### Enhancements